place mirrors that are slow (from our perspective) at the end of the queue
most mirrors are really fast from the host running check-mirror's
perspective, but a couple take a long time to transfer files for signature
verification. as a result, there's a big bandwidth spike at the beginning,
then a flat trough that lasts a couple hours, then a big spike once we get
past the slow mirrors. take the bandwidth hit early so we can schedule this
job sanely; the relative bandwidth trickle from slow mirrors is entirely
acceptable at the end.