dynamically adapt to mirror transfer rates, queueing faster mirrors first
slow_verify involves someone watching this output and caring, neither
of which i have a strong desire to do.

instead, keep track of the amount of time it last took to verify each
mirror's signatures, and check faster mirrors first so we take the
biggest bandwidth spike at the beginning.